Add net8.0 target and AVX-512/Vector512 parsers (#23)
* Use `nuint` for masks and only use Vector512 in 64-bit process.
* Add benchmark results to `benchmarks` directory.
* .NET 8.0 Preview 7 appears to offer nice minor improvements across
everything. Some major for floats.

+++ src/Sep/Internals/SepParserIndexOfAny.cs
@@ -10,7 +10,7 @@ sealed class SepParserIndexOfAny : ISepParser
 {
     readonly char _separator;
     readonly char[] _specialChars;
-    int _quoting = 0;
+    nuint _quoting = 0;
 
     public unsafe SepParserIndexOfAny(Sep sep)
     {
